Remove a few more couchdb configurables
I imagine sometime in the 2020s we'll be able to grep the code base
and not see `couch` in it. I found this when reviewing available
erchef configuration. Grep shows no uses of it. I have not tested this
at all. If something is using this I really want to find out.

Signed-off-by: Steven Danna <steve@chef.io>