[omnibus] Rabbitmq uses loopback for its name
WIP; This doesn't work in all configurations, refining still

localhost is unsafe in some configurations; not all systems resolve it
consistiently. Trying ipv4 loopback (127.0.0.1) as an alternative.

NOTE: Erlang distribution can be either a shortname (myhost) or a
longname (myhost.example.com). IPv4 addresses are treated as
longnames, and we have to switch to using those.

Signed-off-by: Mark Anderson <mark@chef.io>